LayoutTests/fast/dom/htmlcollection-getownpropertynames-expected.txt

Tests that HTMLCollection's named properties returned by Object.getOwnPropertyNames() but not enumerable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Own properties
Actual:   ['0', '1', '2', '3', '4', '5', '6', '7', 'some-id', 'some-name', 'another-id', 'another-name', 'length']
Expected: ['0', '1', '2', '3', '4', '5', '6', '7', 'some-id', 'some-name', 'another-id', 'another-name']

* Enumerated properties
Actual:   ['0', '1', '2', '3', '4', '5', '6', '7', 'length', 'item', 'namedItem']
Expected: ['0', '1', '2', '3', '4', '5', '6', '7', 'length', 'item', 'namedItem']
PASS successfullyParsed is true

TEST COMPLETE
